% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{albania}
\alias{albania}
\title{Albania 2017 Election Results by Polling Station}
\format{A data frame with 5362 rows and 45 variables
\describe{
\item{qarku}{district, 12 in total}
\item{Q_ID}{geocode for district}
\item{bashkia}{municipality, 61 in total}
\item{BAS_ID}{geocode for municipality}
\item{zaz}{election area zone, 90 in total}
\item{njesiaAdministrative}{village, 373 in total}
\item{COM_ID}{geocode for village}
\item{qvKod}{polling station identifier}
\item{zgjedhes}{number of total registered voters}
\item{meshkuj}{number of male registered voters}
\item{femra}{number of female registered voters}
\item{totalSeats}{number of seats contested by district}
\item{vendndodhja}{name of polling center containing polling stations}
\item{ambienti}{type of polling center, 5 in total}
\item{totalVoters}{number of total registered voters that cast ballots}
\item{femVoters}{number of female registered voters that cast ballots}
\item{maleVoters}{number of male registered voters that cast ballots}
\item{unusedBallots}{number of ballots not used}
\item{damagedBallots}{number of ballots damaged}
\item{ballotsCast}{number of total ballots cast}
\item{invalidVotes}{number of ballots cast that were invalidated}
\item{validVotes}{number of valid ballots cast}
\item{lsi}{number of ballots cast for LSI}
\item{ps}{number of ballots cast for PS}
\item{pkd}{number of ballots cast for PKD}
\item{sfida}{number of ballots cast for SFIDA}
\item{pr}{number of ballots cast for PR}
\item{pd}{number of ballots cast for PD}
\item{pbdksh}{number of ballots cast for PBDKSH}
\item{adk}{number of ballots cast for ADK}
\item{psd}{number of ballots cast for PSD}
\item{ad}{number of ballots cast for AD}
\item{frd}{number of ballots cast for FRD}
\item{pds}{number of ballots cast for PDS}
\item{pdiu}{number of ballots cast for PDIU}
\item{aak}{number of ballots cast for AAK}
\item{mega}{number of ballots cast for MEGA}
\item{pksh}{number of ballots cast for PKSH}
\item{apd}{number of ballots cast for APD}
\item{libra}{number of ballots cast for LIBRA}
\item{psSeats}{number of seats won by PS}
\item{pdSeats}{number of seats won by PD}
\item{lsiSeats}{number of seats won by LSI}
\item{pdiuSeats}{number of seats won by PDIU}
\item{psdSeats}{number of seats won by PSD}
}}
\source{
\url{https://albaniaelectiondata.herokuapp.com/}
}
\usage{
albania
}
\description{
Data set containing 2017 Albania election results by polling station published by the Central Election Commission and opened by the Coalition of Domestic Observers & Democracy International.
}
\keyword{datasets}
