% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/astrata.R
\name{astrata}
\alias{astrata}
\title{Stratum allocation}
\usage{
astrata(samplesize, N, method = "mixed", min = 1, wp = 1, e = NULL,
  deff = 1, rr = 1, p = 0.5)
}
\arguments{
\item{samplesize}{Total sample size expected. Not needed when the method used is \code{error}.}

\item{N}{Population size. If the population size is not specified, MOEs are estimated assumming an infinite population.}

\item{method}{Allocation can be assigned using different methods. The default method is \strong{mixed}.
\itemize{
  \item \strong{mixed}: Combines equal and proportional allocation across strata. The weight of the proportional assigment  needs to be defined using \code{wp}. If \code{wp = 1} (default value), the strata allocation is proportional. If \code{wp = 0}, the strata allocation is equal (same size per stratum). If \code{wp} is any number between 0 and 1, the strata allocation will be a combination between equal size and proportionality across strata.
  \item \strong{error}: The sample size of strata is based on a margin of error. \code{e} needs to be specified.
  \item \strong{neyman}: Uses Neyman allocation. It assigns sample units within each stratum proportional to the product of the population stratum size and the within-stratum standard deviation, so that minimum variance for a population mean estimator can be achieved.
  \item \strong{root}: Assigns sample units proportionality using the square root of the population sizes.
  \item \strong{stdev}: It assigns sample units proportional to the standard deviation across strata.
}}

\item{min}{Minimum sample size expected per stratum. Default  value is 1.}

\item{wp}{Weight of the proportional allocation. This is only needed when the method is \code{mixed}. If \code{wp = 1} (default value), the strata allocation will be  completly propotional. If \code{wp = 0}, the strata allocation will be completely equal.}

\item{e}{Margin of error. Only needed if method is \code{error}.}

\item{deff}{Design effect (DEff). Generally, from previous studies or estimations.}

\item{rr}{Value or vector of values with the expected response rate, it migth higher than 0 and lower than 1.}

\item{p}{Value or vector of values with population proportions to be estimated by stratum.}
}
\description{
Allocates cases to strata using different methods.
}
\examples{
astrata(200, c(200, 300, 500))
}

