% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chile}
\alias{chile}
\title{Data from Chile}
\format{A data frame with 15 rows and 3 variables:
\describe{
  \item{reg}{15 administrative regions}
  \item{pob}{population}
 \item{pr}{proportion of a fake variable}
  ...
}}
\source{
\url{http://www.ine.cl/}
}
\usage{
chile
}
\description{
A dataset containing regions of Chile, population, and a fake variable (proportion)
}
\keyword{datasets}

