% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedupe.R
\name{dedupe}
\alias{dedupe}
\title{Removes duplicate observations within collected data}
\usage{
dedupe(df, col_name)
}
\arguments{
\item{df}{object containing data frame of collected data}

\item{col_name}{variable within data frame by which to filter for duplicate values}
}
\value{
Returns table of all data based on unique values within collected data
}
\description{
Removes duplicate observations within collected data
}
\examples{
aldupe <- rsamp(df=albania, n=390, rep=TRUE)
dedupe(df=aldupe, col_name=qvKod)
}
