% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dupe.R
\name{dupe}
\alias{dupe}
\title{Identifies duplicate values within collected data}
\usage{
dupe(df, col_name)
}
\arguments{
\item{df}{object containing data frame of collected data}

\item{col_name}{variable within data frame by which to filter for duplicate values}
}
\value{
Returns table of duplicate values within collected data
}
\description{
Identifies duplicate values within collected data
}
\examples{
aldupe <- rsamp(df=albania, n=390, rep=TRUE)
dupe(df=aldupe, col_name=qvKod)
}
