% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{opening}
\alias{opening}
\title{Albania 2017 CDO Election Observation Data Findings on Opening Process}
\format{A data frame with 524 rows and 19 variables
\describe{
\item{qarku}{district, 12 in total}
\item{psID}{polling station identifier}
\item{votersList}{number of registered voters at the polling station}
\item{ballotPapers}{number of ballot papers at the polling station}
\item{pubPriv}{type of polling station, public or private}
\item{openTime}{time when polling station opening, in 30 minute ranges}
\item{numKommish}{number of commissioners present at polling station}
\item{secrecyOpen}{yes-no if polling station enabled voters to cast ballots in secrecy, po or jo}
\item{movementOpen}{yes-no if polling station provided sufficient space to vote, po or jo}
\item{removeMatInside}{yes-no if campaign materials were removed from inside polling station, po or jo}
\item{removeMatOutside}{yes-no if campaign materials were removed from outside polling station, po or jo}
\item{pvComplete}{yes-no if commissioners completed the opening record checklist sheet, po or jo}
\item{boxChecked}{yes-no if commissioners checked to ensure the ballot box was empty before opening, po or jo}
\item{boxSealed}{yes-no if commissioners sealed the ballot box to prevent ballot tampering, po or jo}
\item{recordBox}{yes-no if commissioners recorded the seal number on the ballot box, po or jo}
\item{centerMat}{yes-no if there were all election materials were available at the polling station, po or jo}
\item{blindTools}{yes-no if the polling station was equipped for blind voters, po or jo}
\item{disabledTools}{yes-no-partially if the polling station was equipped for disabled voters, po or jo or pjeserisht}
\item{overallOpen}{very good-good-problematic-very problematic an overall assessment of the opening process, shummir,mir,meprob,shumprob}
}}
\source{
\url{https://ona.io/cdo/35080/216662}
}
\usage{
opening
}
\description{
Data set containing 2017 Albania election observation findings on polling station opening process by the Coalition of Domestic Observers (CDO)
CDO conducted a statistically-based observation (SBO) exercise, deploying observers to a random sample of polling stations for the 25 June 2017 Albanian elections.
This is a subset of observation data collected by CDO observers that includes data that was used to perform statistical analysis.
}
\keyword{datasets}
