% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package-sampler.R
\docType{package}
\name{sampler}
\alias{sampler}
\alias{sampler-package}
\title{sampler: A Simple Package to Compute Margins of Error (MOE) and Sample Sizes.}
\description{
Functions to estimate MOEs and sample sizes for a proportion with different design options.
}
\details{
\itemize{
  \item Version: 0.01
  \item Date: 2015-04-25
  \item URL: \url{http://github.com/sdaza/sampler}
}
}
\section{Functions}{


\itemize{
  \item \code{\link{ssize}}: Compute sample sizes.
  \item \code{\link{serr}}: Compute margin of error (MOE).
  \item \code{\link{astrata}}: Function to assign sample sizes to strata.
  \item \code{\link{serrst}}: Function to compute margin of error for a stratifed sample.
}
}
\author{
Sebastian Daza \email{sdaza@ssc.wisc.edu}
}

