% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/serr.R
\name{serr}
\alias{serr}
\title{Estimation of margins of errors (MOEs) for a proportion}
\usage{
serr(n, deff = 1, rr = 1, N = NULL, cl = 0.95, p = 0.5,
  relative = FALSE)
}
\arguments{
\item{n}{Size of the sample.}

\item{deff}{Design effect (DEff) provided from previous studies or design estimations (e.g., 1.5).}

\item{rr}{Proportion representing expected response rate, it migth be 0 or more, and less than 1.}

\item{N}{Population size. If the population size is not specified, MOEs are estimated using the infinite population formula.}

\item{cl}{Confidence levels, e.g., .95}

\item{p}{Value or vector of values with the expected proportion of each stratum.}

\item{relative}{Logical parameter. Estimate a relative error if TRUE.}
}
\description{
This function provides flexibility to estimate margins of errors for a proportion assuming different population sizes, confidence levels, design effects (DEff), response rates.
}
\examples{
serr(n = 400, deff = 1.5, rr = .8, N = 1000)
}

