% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/serrst.R
\name{serrst}
\alias{serrst}
\title{Estimation of margins of errors (MOEs) for a proportion in a stratified sample}
\usage{
serrst(n, N = NULL, W = NULL, deff = 1, rr = 1, p = 0.5, cl = 0.95,
  relative = FALSE)
}
\arguments{
\item{n}{Vector of sample size of each stratum.}

\item{N}{Value or vector of values with the population size of each stratum.}

\item{W}{Proportion of each stratum in the target population. If the population size is not specified, MOEs are estimated using the infinite population formula.}

\item{deff}{A value or vector of values with design effects (DEff) provided from previous studies.}

\item{rr}{Proportion representing expected response rate, it migth be 0 or more, and less than 1.}

\item{p}{Value or vector of values with the expected proportion of each stratum.}

\item{cl}{Confidence levels, e.g., .95}

\item{relative}{Logical parameter. Estimate a relative error if TRUE.}
}
\description{
This function provides flexibility to estimate margins of errors for a proportion with a stratified sample assuming different population sizes, confidence levels, design effects (DEff), and response rates.
}
\examples{
serrst(n = c(100, 300, 100), N = c(500, 1000, 600))
}

