% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/astrata.R
\name{astrata}
\alias{astrata}
\title{Stratum allocation}
\usage{
astrata(samplesize, N, method = "mixed", min = 1, wp = 1, e = NULL,
  deff = 1, rr = 1, p = 0.5)
}
\arguments{
\item{samplesize}{Total sample size expected. Not needed when method is error.}

\item{N}{Population size. If the population size is not specified, MOEs are estimated using the infinite population formula.}

\item{method}{The allocation can be assigned using different methods. The default method is \strong{mixed}.
\itemize{
  \item \strong{mixed}: Combine equal size  across strata (equal sample size by stratum) and proportional allociation. It needs to specify the weight of the proportional assignation by \code{wp}. If \code{wp = 1} (default value), the strata allocation is proportional. If \code{wp = 0}, the strata allocation is equal. If \code{wp} is any number between 0 and 1, the strata allocation will be the corresponding combination between equal size and proportionality across strata.
  \item \strong{error}: The sample size of strata is based on a margin of error. \code{e} needs to be specified.
  \item \strong{neyman}: It uses Neyman allocation. It assigns sample units within each stratum proportional to the product of the population stratum size and the within-stratum standard deviation, so that minimum variance for a population mean estimator can be achieved.
  \item \strong{root}: It proportionally assigns sample units using the square root of the population sizes.
  \item \strong{stdev}: It assigns sample units proportional to the standard deviation across strata.
}}

\item{min}{Minimum expected sample size per stratum. Default is 1.}

\item{wp}{Weight given to a proportional allocation. This is only need when method is mixed. If wp = 1 (default value), the strata allocation will be propotional. If wp = 0, the strata allocation will be equal.}

\item{e}{Margin of error. Only needed if method is error.}

\item{deff}{Design effect (DEff) provided from previous estimations (e.g., 1.5).}

\item{rr}{Value or vector of values of the expected response rate, it migth higher than 0 and lower than 1.}

\item{p}{Value or vector of values with the proportion to be estimated by stratum.}
}
\description{
This function allocates sample size to strata using different methods.
}
\examples{
astrata(200, c(200, 300, 500))
}

