% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler-package.R
\docType{package}
\name{sampler-package}
\alias{sampler-package}
\alias{sampler}
\title{Aggregated and overdispersed sampling}
\description{
The sampler package is meant to Generate aggregated or overdispersed sampling
designs for any given phylogenetic tree, spatial points or distance matrix.
Results can be used to design experiments/samples, for resample proposes and
data bias removal.
}
\details{
The package includes 4 functions:
\cr*\code{\link{run_sampler}}: Generate aggregated or overdispersed
sampling designs for any given distance matrix.
\cr*\code{\link{run_sampler_geo}}: Applies the run_sampler algorithm to
 geographic coordinates.
\cr*\code{\link{run_sampler_phy}}: Applies the run_sampler algorithm to
 phylogenetic trees.
\cr*\code{\link{Nee_May_1997}}: Generate a subsample from a phylogenetic
tree that maximizes the remaining amount of evolutionary history
(see Nee and May 1997).
}
\references{
Sean Nee and Robert M. May. "Extinction and the loss of evolutionary history." Science 278.5338 (1997): 692-694.
}
\author{
\strong{Bruno Vilela} (maintainer), \strong{Ty Tuff}, \strong{Trevor Fristoe},
\strong{Edric Gavin}, \strong{Lucas Jardim}, \strong{Sara Varela}
 & \strong{Carlos Botero}
}
