% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampler.R
\name{Sampler}
\alias{Sampler}
\title{Sampler for catch-age from fishery data}
\usage{
Sampler(yr = 2014)
}
\arguments{
\item{yr}{for year of sample construction}
}
\value{
estimates of absolute age compositions
}
\description{
Main code for merging data from obsint.debriefed_age and debriefed_length for sampler code
requires catch, age, and length frequency data in "import" directory with naming convention
catch2014.csv, len2014.csv, age2014.csv
}
\author{
J Ianelli
}

