% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package-sampler.R
\docType{package}
\name{sampler}
\alias{sampler}
\alias{sampler-package}
\title{sampler: A simple package to estimate MOEs and sample sizes.}
\description{
Functions to estimate MOEs and sample sizes for a proportion includig different design options.
}
\details{
\itemize{
  \item Version: 0.01
  \item Date: 2015-04-24
  \item URL: \url{http://github.com/sdaza/sampler}
}
}
\section{Functions}{


\itemize{
  \item \code{\link{ssize}}: Compute sample sizes.
  \item \code{\link{serr}}: Compute margin of error (MOE).
  \item \code{\link{astrata}}: Function to assign strata's sample sizes using different procedures.
  \item \code{\link{serrst}}: Function to compute margin of error for a stratifed sample.
  \item \code{\link{serrst}}: Function to compute margin of error for a stratifed sample.
}
}
\author{
Sebastian Daza \email{sdaza@ssc.wisc.edu}
}

